mac osx 10.4 version, tested on mpi codes only
git-svn-id: https://svn.mcs.anl.gov/repos/osuflow/trunk@77 6b533941-e25a-0410-8d37-8675ae1e3315

+++ FlashAMR.C
@@ -1,5 +1,5 @@
 
-#include "flashhdf5_float.h"
+// #include "flashhdf5_float.h"
 #include "FlashAMR.h" 
 
 void sort_list(float* list, int& cnt) 
@@ -12,7 +12,7 @@ void sort_list(float* list, int& cnt)
     for (int j = i; j<cnt; j++) {
       if (tmp_list[j]>largest) {idx = j; largest = tmp_list[j]; }
     }
-    int tmp = tmp_list[i]; 
+    int tmp = (int)tmp_list[i]; 
     tmp_list[i] = tmp_list[idx]; 
     tmp_list[idx] =tmp; 
   }
